requirements.txt

asgiref==3.6.0
backports.zoneinfo==0.2.1
Django~=4.1.3
django-autoslug==1.9.8
django-sass-processor==1.2.2
Pillow==9.4.0
pip==23.0.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
setuptools==49.2.1
sqlparse==0.4.3
django_debug_toolbar==3.8.1

django-debug-toolbar~=3.8.1
djangorestframework~=3.14.0